ci: Generate Fluent journals
