+++ .github/workflows/test-fluent-journals.yml
@@ -82,5 +82,4 @@ jobs:
 
       - name: Run Fluent tests
         run: |
-          python -m pytest --write-fluent-journals
-          python .ci/fluent_test_runner.py tests
+          make write-and-run-fluent-tests
