+++ tests/test_launcher_remote.py
@@ -2,6 +2,8 @@
 
 from concurrent import futures
 import os
+from pathlib import Path
+import shutil
 from unittest.mock import create_autospec
 import uuid
 
@@ -12,7 +14,7 @@ from test_session import MockHealthServicer, MockSchemeEvalServicer
 
 from ansys.api.fluent.v0 import scheme_eval_pb2_grpc
 import ansys.fluent.core as pyfluent
-from ansys.fluent.core import examples
+from ansys.fluent.core import EXAMPLES_PATH, examples
 from ansys.fluent.core.fluent_connection import (
     FluentConnection,
     UnsupportedRemoteFluentInstance,
@@ -26,7 +28,6 @@ import ansys.fluent.core.utils.fluent_version as docker_image_version
 from ansys.fluent.core.utils.fluent_version import FluentVersion
 from ansys.fluent.core.utils.networking import get_free_port
 import ansys.platform.instancemanagement as pypim
-from tests.util import rename_downloaded_file
 
 
 def test_launch_remote_instance(monkeypatch, new_solver_session):
@@ -141,6 +142,81 @@ class TransferRequestRecorder:
 
     def is_configured(self):
         return True
+
+
+def rename_downloaded_file(file_path: str, suffix: str) -> str:
+    """Rename downloaded file by appending a suffix to the file name.
+
+    Parameters
+    ----------
+    file_path : str
+        Downloaded file path. Can be absolute or relative.
+    suffix : str
+        Suffix to append to the file name.
+
+    Returns:
+    --------
+    str
+        New file path with the suffix appended to the file name.
+    """
+    ext = "".join(Path(file_path).suffixes)
+    orig_path = Path(file_path)
+    file_path = file_path.removesuffix(ext)
+    file_path = Path(file_path)
+    if file_path.is_absolute():
+        new_stem = f"{file_path.stem}{suffix}"
+        new_path = file_path.with_stem(new_stem)
+        new_path = new_path.with_suffix(ext)
+        orig_path.rename(new_path)
+        return str(new_path)
+    else:
+        orig_abs_path = Path(EXAMPLES_PATH) / orig_path
+        abs_path = Path(EXAMPLES_PATH) / file_path
+        new_stem = f"{file_path.stem}{suffix}"
+        new_path = abs_path.with_stem(new_stem)
+        new_path = new_path.with_suffix(ext)
+        orig_abs_path.rename(new_path)
+        return str(file_path.with_stem(new_stem).with_suffix(ext))
+
+
+@pytest.mark.parametrize(
+    "ext,a,b,c,d",
+    [(".cas", "a1", "b1", "c1", "d1"), (".cas.gz", "a2", "b2", "c2", "d2")],
+)
+def test_rename_downloaded_file(ext, a, b, c, d):
+    try:
+        file_path = Path(EXAMPLES_PATH) / f"{a}{ext}"
+        file_path.touch()
+        file_path = str(file_path)
+        new_file_path = rename_downloaded_file(file_path, "_1")
+        assert new_file_path == str(Path(EXAMPLES_PATH) / f"{a}_1{ext}")
+    except Exception:
+        raise
+    finally:
+        Path(new_file_path).unlink(missing_ok=True)
+
+    try:
+        file_path = f"{b}{ext}"
+        (Path(EXAMPLES_PATH) / file_path).touch()
+        new_file_path = rename_downloaded_file(file_path, "_1")
+        assert new_file_path == f"{b}_1{ext}"
+    except Exception:
+        raise
+    finally:
+        (Path(EXAMPLES_PATH) / new_file_path).unlink(missing_ok=True)
+
+    try:
+        dir_path = Path(EXAMPLES_PATH) / c
+        dir_path.mkdir()
+        file_path = dir_path / f"{d}{ext}"
+        file_path.touch()
+        file_path = str(Path(c) / f"{d}{ext}")
+        new_file_path = rename_downloaded_file(file_path, "_1")
+        assert new_file_path == str(Path(c) / f"{d}_1{ext}")
+    except Exception:
+        raise
+    finally:
+        shutil.rmtree(dir_path, ignore_errors=True)
 
 
 @pytest.mark.codegen_required
